Migrate annual users report generation
